Debounces URL parameter updates.
There is a security feature in safari that prevents an application from
using the history.pushState() function more than 100 times in 30
seconds. If that trigger is reached, the page reloads and messes up our
rendering. Adding a debounce drastically reduces the number of times the
url is updated if a slider is "wiggled".