Prevent multiple clicks of mirror button
When the mirror button is clicked, it gains focus. Once a button has
focus, pressing the spacebar will trigger a further click event on that
button. We are using the spacebar to toggle the EM skeleton. To prevent
the mirror button from being clicked when the EM skeleton is toggled, the
target <span> and the parent <button> were blurred after a click to
prevent continued clicks. This fixes #4