Parameter updates replace instead of add to history
Before this update any parameter change would lead to a new entry in the
browser history. This was annoying if the user wanted to go back to the
site they had come from by hitting the back button. Changing the
parameter update from a push to a replace, means they can get back to
the referencing site with one click of the back button, instead of many.